functions/requirements.txt
autopep8==2.0.0
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.3
cloudevents==1.9.0
cloudinary==1.30.0
cryptography==41.0.7
dataclasses-json==0.6.3
deprecation==2.1.0
firebase-admin==6.3.0
firebase-functions==0.1.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.2
functions-framework==3.5.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-events==0.11.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==2.0.1
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.20.1
msgpack==1.0.7
mypy-extensions==1.0.0
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sentry-sdk==1.39.1
six==1.16.0
SQLAlchemy==1.4.45
tomli==2.0.1
typing-inspect==0.9.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.13
watchdog==3.0.0
Werkzeug==3.0.1
